+++ src/npm-client/install-packages.js
@@ -57,18 +57,33 @@ export async function installNpmPackagesForComponents(
   componentsWithDependencies: ComponentWithDependencies[],
   verbose: boolean = false,
   silentPackageManagerResult: boolean = false,
-  installPeerDependencies: boolean = false
+  installPeerDependencies: boolean = false,
+  hasImportedComponentsAddedToRootPackageJson: boolean = true
 ): Promise<*> {
-  // if dependencies are installed as bit-components, go to each one of the dependencies and install npm packages
-  // otherwise, if the dependencies are installed as npm packages, npm already takes care of that
+  const installImportedPackagesViaRoot =
+    (await npmClient.isSupportedInstallationOfSubDirFromRoot(consumer.bitJson.packageManager)) &&
+    hasImportedComponentsAddedToRootPackageJson;
+
   const componentsWithDependenciesFlatten = R.flatten(
     componentsWithDependencies.map((oneComponentWithDependencies) => {
-      return oneComponentWithDependencies.component.dependenciesSavedAsComponents
-        ? [oneComponentWithDependencies.component, ...oneComponentWithDependencies.dependencies]
+      const componentsToInstallPackages = installImportedPackagesViaRoot
+        ? [] // no need to install the imported components in their subDir, it'd be installed later from the rootDir
         : [oneComponentWithDependencies.component];
+      // if dependencies are installed as bit-components, go to each one of the dependencies and install npm packages
+      // otherwise, if the dependencies are installed as npm packages, npm already takes care of that
+      if (oneComponentWithDependencies.component.dependenciesSavedAsComponents) {
+        componentsToInstallPackages.push(...oneComponentWithDependencies.dependencies);
+      }
+      return componentsToInstallPackages;
     })
   );
 
   const componentDirs = componentsWithDependenciesFlatten.map(component => component.writtenPath);
-  return installPackages(consumer, componentDirs, verbose, false, silentPackageManagerResult, installPeerDependencies);
+  await installPackages(consumer, componentDirs, verbose, false, silentPackageManagerResult, installPeerDependencies);
+
+  if (installImportedPackagesViaRoot) {
+    // install package.json of the rootDir, which will install the packages of the imported
+    // components in a smart way, where the shared packages are hoisted to root dir
+    await installPackages(consumer, [], verbose, true, silentPackageManagerResult, installPeerDependencies);
+  }
 }